прикрутил фильтр

+++ src/main/java/com/space/repository/ShipRepository.java
@@ -3,7 +3,7 @@ package com.space.repository;
 import com.space.model.Ship;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ShipRepository extends CrudRepository<Ship, Integer> {
+public interface ShipRepository extends CrudRepository<Ship, Long> {
 
 }
